Split runnner.py into bootstrap.py and runner.py.
Tests should pull runner.py from the repository and not the image.
