+++ test-infra/runner.py
@@ -57,64 +57,6 @@ def run(command, cwd=None):
   logging.info("Running: %s", " ".join(command))
   subprocess.check_call(command, cwd=cwd)
 
-def clone_repo():
-  """Clone the repo.
-
-  Returns:
-    src_path: This is the root path for the training code.
-    sha: The sha number of the repo.
-  """
-  go_path = os.getenv("GOPATH")
-  # REPO_OWNER and REPO_NAME are the environment variables set by Prow.
-  repo_owner = os.getenv("REPO_OWNER")
-  repo_name = os.getenv("REPO_NAME")
-
-  if bool(repo_owner) != bool(repo_name):
-    raise ValueError("Either set both environment variables "
-                     "REPO_OWNER and REPO_NAME or set neither.")
-
-  src_dir = os.path.join(go_path, "src/github.com", GO_REPO_OWNER)
-  dest = os.path.join(src_dir, GO_REPO_NAME)
-
-  if not repo_owner and not repo_name:
-    logging.info("Environment variables REPO_OWNER and REPO_NAME not set; "
-                 "not checking out code.")
-    if not os.path.exists(dest):
-      raise ValueError("No code found at %s", dest)
-    # TODO(jlewi): We should get the sha number and add "-dirty" if needed
-    return dest, ""
-
-  # Clone mlkube
-  repo = "https://github.com/{0}/{1}.git".format(repo_owner, repo_name)
-  logging.info("repo %s", repo)
-
-  os.makedirs(src_dir)
-
-  # TODO(jlewi): How can we figure out what branch
-  run(["git", "clone",  repo, dest])
-
-  # If this is a presubmit PULL_PULL_SHA will be set
-  # see:
-  # https://github.com/kubernetes/test-infra/tree/master/prow#job-evironment-variables
-  sha = os.getenv('PULL_PULL_SHA')
-
-  if not sha:
-    # For postsubmits PULL_BASE_SHA will be set.
-    sha = os.getenv('PULL_BASE_SHA')
-
-  if sha:
-    run(["git", "checkout", sha], cwd=dest)
-
-  # Install dependencies
-  run(["glide", "install"], cwd=dest)
-
-  # We need to remove the vendored dependencies of apiextensions otherwise we
-  # get type conflicts.
-  shutil.rmtree(os.path.join(dest,
-                             "vendor/k8s.io/apiextensions-apiserver/vendor"))
-
-  return dest, sha
-
 class TimeoutError(Exception):
   """An error indicating an operation timed out."""
 
@@ -426,6 +368,17 @@ if __name__ == "__main__":
       type=str,
       help="Zone to use for spinning up the GKE cluster.")
 
+  parser.add_argument(
+    "--src_dir",
+    default="",
+      type=str,
+      help="The source directory.")
+
+  parser.add_argument(
+    "--sha",
+    default="",
+      type=str,
+      help="The sha of the code.")
 
   parser.add_argument("--gcb", dest="use_gcb", action="store_true",
                       help="Use Google Container Builder to build the image.")
@@ -442,6 +395,12 @@ if __name__ == "__main__":
   parser.set_defaults(use_gke=True)
 
   args = parser.parse_args()
+
+  if not args.src_dir:
+    raise ValueError("--src_dir must be specified.")
+
+  src_dir = args.src_dir
+  sha = args.sha
 
   # Print environment variables.
   # This is useful for debugging and understanding the information set by prow.
@@ -462,8 +421,6 @@ if __name__ == "__main__":
                  os.getenv("GOOGLE_APPLICATION_CREDENTIALS"))
     run(["gcloud", "auth", "activate-service-account",
          "--key-file={0}".format(os.getenv("GOOGLE_APPLICATION_CREDENTIALS"))])
-
-  src_dir, sha = clone_repo()
 
   output_dir = get_gcs_output()
   gcs_client = storage.Client()
